Use MouseMovementCallback for camera manipulation.
This is is more direct, and there's no need for us to calculate the
delta mouse movement ourselves (goglfw does it for us). This works
better when cursor is disabled or when cursor is at the edge of screen
because it continues to provide true mouse movement deltas (potentially
without pointer ballistics).
This is an investigation of an experimental API addition related to
glfw/glfw#125; if it goes well I'll make a PR for glfw. /cc @elmindreda