Pretend Go has generics to better solve this Java problem.
Thankfully, I don't have to do this in normal idiomatic Go code, which
makes it clean, easy to read and reason about, because it's more
concrete.
Note that unicode characters that look like angle brackets are used,
not actual angle brackets (because those are reserved keyword for
comparison operators). But Go allows unicode characters in its
identifier names, so it works and compiles (despite being awful). Don't
actually do this unless your task is to write Java code!